Moved from_dict and as_dict methods into parent Model class
